requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
cytoolz==0.12.0
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-loading-spinners==1.0.0
dash-table==5.0.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Delorean==1.0.0
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Compress==1.13
fonttools==4.37.3
frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
gunicorn==20.1.0
hexbytes==0.3.0
httplib2==0.21.0
hubspot==0.1.14.dev0
hubspot-api-client==7.1.0
humanize==4.4.0
idna==3.4
importlib-metadata==5.1.0
ipfshttpclient==0.8.0a2
ipinfo==4.4.2
ipykernel==6.17.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
kaleido==0.2.1
kiwisolver==1.4.4
lru-dict==1.1.8
lxml==4.9.1
mailjet-rest==1.3.4
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
multiaddr==0.0.9
multidict==6.0.3
multitasking==0.0.11
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.0
plotly==5.11.0
prompt-toolkit==3.0.36
protobuf==3.19.5
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.16.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.28.1
rlp==2.0.1
rsa==4.9
scipy==1.9.3
sendgrid==6.9.7
simplejson==3.18.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
stack-data==0.6.2
starkbank-ecdsa==2.2.0
statsmodels==0.13.5
stripe==5.0.0
tenacity==8.1.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
traitlets==5.6.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
varint==1.0.2
wcwidth==0.2.5
web3==5.31.3
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.2
yarl==1.8.2
yfinance==0.1.90
zipp==3.11.0
